Define a proxy for plugin installation
Address the issue [ansible-elasticsearch
#4](https://github.com/elastic/ansible-elasticsearch/issues/4) which
add proxy for plugin installation and more generally es_proxy_host and
es_proxy_port for global usage.

Right now, it’s only used for plugins installation. In the futur, it
can be used (or defined) in the java_opts within the default file for
instance (elasticsearch.j2)