Ensure elasticsearch is always started
Ensure that elasticsearch is started/running even if non of the handlers
are called.  Previously, if the host was provisioned correctly, but
elasticsearch was not running the role would not attempt to start it.
This can happen if elasticsearch was killed, or if the previous ansible
run failed prior to running the handlers.